hyggebo_brain/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
aiohttp==3.11.18
paho-mqtt==2.1.0
asyncpg==0.30.0
